+++ resources/scripts/install/install-pocs.sh
@@ -22,7 +22,7 @@ cd "$PANDIR"
 echo "Creating POCS directories."
 mkdir -p "${HOME}/logs"
 mkdir -p "${HOME}/images"
-mkdir -p "${HOME}/json_store"
+mkdir -p "${HOME}/telemetry"
 mkdir -p "${HOME}/keys"
 
 # Link the needed POCS folders.